+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### [Unreleased]
 
+* ADDED: Code editor to full screen button
+* ADDED: Web Bluetooth not supported notification
+
 ### [0.8.0] - 2019-05-12
 
 - ADDED: iOS not supported notification