+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### [Unreleased]
 
+### [0.9.0] - 2019-08-11
+
 - ADDED: Code editor to full size button
 - CHANGED: Minor UI updates
 - ADDED: Web Bluetooth not supported notification